requirements.txt
aerich==0.9.1
aiogram==3.21.0
alembic==1.16.4
asyncpg==0.30.0
fastapi==0.116.1
pydantic-settings==2.10.1
tortoise-orm==0.25.1
uvicorn==0.35.0
